Create base widget class
